Frontend/i18n: Extract i18n date.ts functions to i18n/datetimes.ts
