+++ app/web/components/Datepicker.tsx
@@ -6,9 +6,9 @@ import {
   usePickerContext,
 } from "@mui/x-date-pickers";
 import { useTranslation } from "i18n";
+import { getMuiDateFormat } from "i18n/datetimes";
 import { Control, Controller, UseControllerProps } from "react-hook-form";
 import { Temporal } from "temporal-polyfill";
-import { getMuiDateFormat } from "utils/date";
 import dayjs, { Dayjs } from "utils/dayjs";
 
 interface DatepickerProps {
