+++ app/web/features/communities/discussions/DiscussionPage.tsx
@@ -17,6 +17,7 @@ import { discussionKey } from "features/queryKeys";
 import { useLiteUser } from "features/userQueries/useLiteUsers";
 import { RpcError } from "grpc-web";
 import { useTranslation } from "i18n";
+import { localizeDateTime, localizeRelativeTime } from "i18n/datetimes";
 import { COMMUNITIES, GLOBAL } from "i18n/namespaces";
 import { useRouter } from "next/router";
 import { Discussion } from "proto/discussions_pb";
@@ -24,11 +25,7 @@ import { useState } from "react";
 import { useForm } from "react-hook-form";
 import { service } from "service";
 import { theme } from "theme";
-import {
-  localizeDateTime,
-  localizeRelativeTime,
-  timestampToPlainDateTime,
-} from "utils/date";
+import { timestampToPlainDateTime } from "utils/date";
 
 import { sendNativeBack, useIsNativeEmbed } from "../../../utils/nativeLink";
 import CommunityBase from "../CommunityBase";
