+++ app/web/features/dashboard/DiscussionListRow.tsx
@@ -7,11 +7,11 @@ import {
 import { Skeleton, styled } from "@mui/material";
 import getContentSummary from "features/communities/getContentSummary";
 import { useTranslation } from "i18n";
+import { localizeRelativeTime } from "i18n/datetimes";
 import { DASHBOARD, GLOBAL } from "i18n/namespaces";
 import Link from "next/link";
 import { Discussion } from "proto/discussions_pb";
 import { routeToDiscussion } from "routes";
-import { localizeRelativeTime } from "utils/date";
 
 type DiscussionSummary = Pick<
   Discussion.AsObject,
