+++ app/web/features/messages/requests/HostRequestListItem.tsx
@@ -26,13 +26,13 @@ import useCurrentUser from "features/userQueries/useCurrentUser";
 import { useLiteUser } from "features/userQueries/useLiteUsers";
 import { RpcError } from "grpc-web";
 import { useTranslation } from "i18n";
+import { localizeDateTimeRange } from "i18n/datetimes";
 import { MESSAGES } from "i18n/namespaces";
 import { HostRequest } from "proto/requests_pb";
 import React, { useState } from "react";
 import { service } from "service";
 import { Temporal } from "temporal-polyfill";
 import { theme } from "theme";
-import { localizeDateTimeRange } from "utils/date";
 import { firstName } from "utils/names";
 
 import HostRequestStatusText from "./HostRequestStatusText";
